clarity/requirements.txt
flask-sqlalchemy
flask_login
uWSGI
mysqlclient